board/user 권한설정

+++ src/main/java/shop/mtcoding/blog/controller/BoardController.java
@@ -8,6 +8,8 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import shop.mtcoding.blog._core.errors.exception.Exception403;
+import shop.mtcoding.blog._core.errors.exception.Exception404;
 import shop.mtcoding.blog.board.Board;
 import shop.mtcoding.blog.board.BoardRepository;
 import shop.mtcoding.blog.board.BoardRequest;
@@ -52,17 +54,28 @@ public class BoardController {
         Board board = boardRepository.findByIdJoinUser(id);
 
 
+        boolean isOwner = false;
+        if (user != null){
+            isOwner = false;
 
+            if (user.getId() == board.getUser().getId()){
+                isOwner = true;
+            }
+        }
 
-
+        request.setAttribute("isOwner",isOwner);
         request.setAttribute("board",board);
         return "/board/detail";
     }
 
     @GetMapping("/board/{id}/update-form")
     public String updateform(@PathVariable Integer id, HttpServletRequest request){
-
         Board board = boardRepository.findByIdJoinUser(id);
+
+        if (board.getUser().getId() == null){
+            throw new Exception404("해당 게시글을 찾을 수 없습니다");
+        }
+
         request.setAttribute("board",board);
 
          return "/board/update-form";
@@ -72,7 +85,13 @@ public class BoardController {
     @PostMapping("/board/{id}/update")
     public String update(@PathVariable Integer id,BoardRequest.UpdateDTO reqDTO){
 
+        User user = (User) session.getAttribute("sessionUser");
         Board board = boardRepository.findByIdJoinUser(id);
+
+        if (user.getId() != board.getUser().getId()){
+            throw new Exception403("권한이없습니다.");
+        }
+
         board.update(reqDTO);
 
         return "redirect:/board/"+id;
@@ -80,6 +99,12 @@ public class BoardController {
 
     @PostMapping("/board/{id}/delete")
     public String delete(@PathVariable Integer id){
+        User user = (User) session.getAttribute("sessionUser");
+        Board board = boardRepository.findByIdJoinUser(id);
+
+        if (user.getId() != board.getUser().getId()){
+            throw new Exception403("권한이없습니다.");
+        }
 
         boardRepository.deleteById(id);
